+++ src/audio/dai.c
@@ -6,7 +6,7 @@
 //         Keyon Jie <yang.jie@linux.intel.com>
 
 #include <sof/audio/buffer.h>
-#include <sof/audio/component.h>
+#include <sof/audio/component_ext.h>
 #include <sof/audio/format.h>
 #include <sof/audio/pcm_converter.h>
 #include <sof/audio/pipeline.h>
