+++ src/audio/kpb.c
@@ -15,7 +15,7 @@
  */
 
 #include <sof/audio/buffer.h>
-#include <sof/audio/component.h>
+#include <sof/audio/component_ext.h>
 #include <sof/audio/pipeline.h>
 #include <sof/audio/kpb.h>
 #include <sof/common.h>
